Refactor imports by removing index.js files
