+++ assets/react/controllers/Compontents/Button/Button.jsx
@@ -8,10 +8,8 @@ import './index.css';
  * @returns {JSX.Element}
  * @constructor
  */
-function Button({children, onClick}) {
+export default function Button({children, onClick}) {
     return (
         <button className="button" onClick={onClick}>{children}</button>
     );
 }
-
-export default Button;
