--- assets/react/controllers/Compontents/Button/index.js
@@ -1,1 +1,0 @@
-export * from './Button';
