--- assets/react/controllers/Compontents/TaskList/index.js
@@ -1,1 +1,0 @@
-export * from './TaskList';
