--- assets/react/controllers/Compontents/TaskListContainer/index.js
@@ -1,1 +1,0 @@
-export * from './TaskListContainer';
