--- assets/react/controllers/Compontents/TaskListForm/index.js
@@ -1,1 +1,0 @@
-export * from './TaskListForm';
