--- assets/react/controllers/Compontents/TaskListFormContainer/index.js
@@ -1,1 +1,0 @@
-export * from './TaskListFormContainer';
